Struct openrr_client::RobotClient

pub struct RobotClient<L, M, N>where
    L: Localization,
    M: MoveBase,
    N: Navigation,{ /* private fields */ }

Implementations

impl<L, M, N> RobotClient<L, M, N>where
L: Localization,
M: MoveBase,
N: Navigation,

pub fn new(
config: OpenrrClientsConfig,
raw_joint_trajectory_clients: HashMap<String, Arc<dyn JointTrajectoryClient>>,
speakers: HashMap<String, Arc<dyn Speaker>>,
localization: Option<L>,
move_base: Option<M>,
navigation: Option<N>
) -> Result<Self, Error>

pub fn set_raw_clients_joint_positions_to_full_chain_for_collision_checker(
&self
) -> Result<(), Error>

Set the current joint positions to the robot kinematic model.

Returns Error::FullChainNotFound when self.full_chain_for_collision_checker is None.

pub fn is_raw_joint_trajectory_client(&self, name: &str) -> bool

pub fn is_joint_trajectory_client(&self, name: &str) -> bool

pub fn is_collision_avoidance_client(&self, name: &str) -> bool

pub fn is_collision_check_client(&self, name: &str) -> bool

pub fn is_ik_client(&self, name: &str) -> bool

pub fn joint_trajectory_clients(
&self
) -> &HashMap<String, Arc<dyn JointTrajectoryClient>>

pub fn self_collision_checkers(
&self
) -> &HashMap<String, Arc<SelfCollisionChecker<f64>>>

pub fn ik_solvers(&self) -> &HashMap<String, Arc<IkSolverWithChain>>

pub fn collision_avoidance_clients(
&self
) -> &HashMap<String, Arc<CollisionAvoidanceClient<Arc<dyn JointTrajectoryClient>>>>

pub fn ik_clients(
&self
) -> &HashMap<String, Arc<IkClient<Arc<dyn JointTrajectoryClient>>>>

pub fn send_joint_positions(
&self,
name: &str,
positions: &[f64],
duration_sec: f64
) -> Result<WaitFuture, Error>

pub fn current_joint_positions(&self, name: &str) -> Result<Vec<f64>, Error>

pub fn joint_names(&self, name: &str) -> Result<Vec<String>, Error>

Returns the joint names which is handled by the joint trajectory client with name

pub fn send_joints_pose(
&self,
name: &str,
pose_name: &str,
duration_sec: f64
) -> Result<WaitFuture, Error>

pub fn current_end_transform(&self, name: &str) -> Result<Isometry3<f64>, Error>

pub fn transform(
&self,
name: &str,
pose: &Isometry3<f64>
) -> Result<Isometry3<f64>, Error>

pub fn move_ik(
&self,
name: &str,
target_pose: &Isometry3<f64>,
duration_sec: f64
) -> Result<WaitFuture, Error>

pub fn move_ik_with_interpolation(
&self,
name: &str,
target_pose: &Isometry3<f64>,
duration_sec: f64,
max_resolution: f64,
min_number_of_points: i32
) -> Result<WaitFuture, Error>

pub fn send_joint_positions_with_pose_interpolation(
&self,
name: &str,
positions: &[f64],
duration_sec: f64,
max_resolution: f64,
min_number_of_points: i32
) -> Result<WaitFuture, Error>

pub fn raw_joint_trajectory_clients_names(&self) -> Vec<String>

pub fn joint_trajectory_clients_names(&self) -> Vec<String>

pub fn collision_avoidance_clients_names(&self) -> Vec<String>

pub fn collision_check_clients_names(&self) -> Vec<String>

pub fn ik_clients_names(&self) -> Vec<String>

pub fn full_chain_for_collision_checker(&self) -> &Option<Arc<Chain<f64>>>

pub fn speakers(&self) -> &HashMap<String, Arc<dyn Speaker>>

pub fn speak(&self, name: &str, message: &str) -> Result<WaitFuture, Error>